feat/pong-together/frontend#12: oauth 로그인 시 상태관리
