+++ src/pages/components/login/OauthBtn.js
@@ -13,7 +13,7 @@ export default class extends Component {
 	setEvent() {
 		this.addEvent('click', '#login-oauth-btn', () => {
 			window.location.href =
-				'https://api.intra.42.fr/oauth/authorize?client_id=u-s4t2ud-9faf7925c4cf5dfe9518b96183f6225ad7c1c45e2f186121a5d808d8dcdad924&redirect_uri=https://localhost:443&response_type=code';
+				'https://api.intra.42.fr/oauth/authorize?client_id=u-s4t2ud-9faf7925c4cf5dfe9518b96183f6225ad7c1c45e2f186121a5d808d8dcdad924&redirect_uri=http://localhost:8080&response_type=code';
 		});
 	}
 
@@ -21,22 +21,5 @@ export default class extends Component {
 		return `<button class="login-btn" id="login-oauth-btn">${language.login[this.$store.state.language].loginBtn}</button>`;
 	}
 
-	async mounted() {
-		const queryParams = new URLSearchParams(window.location.search);
-		const code = queryParams.get('code');
-		console.log(code);
-		if (code) {
-			console.log('요청');
-			try {
-				const data = await http.post(
-					'http://localhost:8000/api/auth/login/',
-					{ code: code },
-					{ 'Content-Type': 'application/json' },
-				);
-				console.log('data', data);
-			} catch (error) {
-				console.error('HTTP 요청 실패:', error);
-			}
-		}
-	}
+	async mounted() {}
 }
